+++ packages/validasi_ui/example/lib/user.g.dart
@@ -7,7 +7,9 @@ sealed class UserFields<V> extends ValidasiKey<User>
   const UserFields._();
 
   static const UserFields<String> name = UserNameField();
+
   static const UserFields<String> email = UserEmailField();
+
   static const UserFields<int> age = UserAgeField();
 }
 
@@ -15,15 +17,18 @@ class UserNameField extends UserFields<String> {
   const UserNameField() : super._();
 
   @override
-  String get name => 'name';
+  String get name {
+    return 'name';
+  }
 
   @override
-  String extract(User owner) => owner.name;
+  String extract(User owner) {
+    return owner.name;
+  }
 
   @override
   ValidasiResult<String> validate(String? value) {
     final $errors = <ValidationError>[];
-
     if (value != null && value.length < 2) {
       $errors.add(
         ValidationError(
@@ -34,7 +39,6 @@ class UserNameField extends UserFields<String> {
         ),
       );
     }
-
     if (value != null && value.length > 100) {
       $errors.add(
         ValidationError(
@@ -52,30 +56,27 @@ class UserNameField extends UserFields<String> {
   }
 
   @override
-  CrossFieldKey<User>? get crossFieldKey => null;
-
-  @override
-  List<ValidationError> Function(V? Function<V>(ValidasiField<User, V>))?
-      get crossValidator => null;
-
-  @override
-  Set<ValidasiField<User, dynamic>> get crossDependsOn =>
-      const <ValidasiField<User, dynamic>>{};
+  Future<ValidasiResult<String>> validateAsync(String? value) async {
+    return validate(value);
+  }
 }
 
 class UserEmailField extends UserFields<String> {
   const UserEmailField() : super._();
 
   @override
-  String get name => 'email';
+  String get name {
+    return 'email';
+  }
 
   @override
-  String extract(User owner) => owner.email;
+  String extract(User owner) {
+    return owner.email;
+  }
 
   @override
   ValidasiResult<String> validate(String? value) {
     final $errors = <ValidationError>[];
-
     if (value != null && value.length < 3) {
       $errors.add(
         ValidationError(
@@ -86,7 +87,6 @@ class UserEmailField extends UserFields<String> {
         ),
       );
     }
-
     if (value != null && value.length > 100) {
       $errors.add(
         ValidationError(
@@ -104,26 +104,8 @@ class UserEmailField extends UserFields<String> {
   }
 
   @override
-  CrossFieldKey<User> get crossFieldKey => UserCrossFields.email;
-
-  @override
-  Set<ValidasiField<User, dynamic>> get crossDependsOn {
-    return const <ValidasiField<User, dynamic>>{UserFields.name};
-  }
-
-  @override
-  List<ValidationError> Function(V? Function<V>(ValidasiField<User, V>))
-      get crossValidator {
-    return (getField) {
-      final result = _emailMatchesName(getField);
-      if (result != null) {
-        return [
-          ValidationError(
-              rule: 'ValidateWith', message: result, path: ['email'])
-        ];
-      }
-      return [];
-    };
+  Future<ValidasiResult<String>> validateAsync(String? value) async {
+    return validate(value);
   }
 }
 
@@ -131,15 +113,18 @@ class UserAgeField extends UserFields<int> {
   const UserAgeField() : super._();
 
   @override
-  String get name => 'age';
+  String get name {
+    return 'age';
+  }
 
   @override
-  int extract(User owner) => owner.age;
+  int extract(User owner) {
+    return owner.age;
+  }
 
   @override
   ValidasiResult<int> validate(int? value) {
     final $errors = <ValidationError>[];
-
     if (value != null && value.length < 1) {
       $errors.add(
         ValidationError(
@@ -157,38 +142,23 @@ class UserAgeField extends UserFields<int> {
   }
 
   @override
-  CrossFieldKey<User>? get crossFieldKey => null;
-
-  @override
-  List<ValidationError> Function(V? Function<V>(ValidasiField<User, V>))?
-      get crossValidator => null;
-
-  @override
-  Set<ValidasiField<User, dynamic>> get crossDependsOn =>
-      const <ValidasiField<User, dynamic>>{};
+  Future<ValidasiResult<int>> validateAsync(int? value) async {
+    return validate(value);
+  }
 }
 
-User assemble_User(ValidasiFormController<User> ctrl) => User(
-      name: ctrl.getValue(UserFields.name) as String,
-      email: ctrl.getValue(UserFields.email) as String,
-      age: ctrl.getValue(UserFields.age) as int,
-    );
-
-sealed class UserCrossFields extends CrossFieldKey<User> {
-  const UserCrossFields._(super.name);
-  static const UserCrossFields email = _User_email_CrossField();
-}
-
-class _User_email_CrossField extends UserCrossFields {
-  const _User_email_CrossField() : super._('email');
+User assemble_User(ValidasiFormController<User> ctrl) {
+  return User(
+    name: ctrl.getValue(UserFields.name) as String,
+    email: ctrl.getValue(UserFields.email) as String,
+    age: ctrl.getValue(UserFields.age) as int,
+  );
 }
 
 extension $UserValidasi on User {
   ValidasiResult<User> validate() {
     final $errors = <ValidationError>[];
-
     // Field: name
-
     if (name != null && name.length < 2) {
       $errors.add(
         ValidationError(
@@ -199,7 +169,6 @@ extension $UserValidasi on User {
         ),
       );
     }
-
     if (name != null && name.length > 100) {
       $errors.add(
         ValidationError(
@@ -210,9 +179,7 @@ extension $UserValidasi on User {
         ),
       );
     }
-
     // Field: email
-
     if (email != null && email.length < 3) {
       $errors.add(
         ValidationError(
@@ -223,7 +190,6 @@ extension $UserValidasi on User {
         ),
       );
     }
-
     if (email != null && email.length > 100) {
       $errors.add(
         ValidationError(
@@ -234,9 +200,7 @@ extension $UserValidasi on User {
         ),
       );
     }
-
     // Field: age
-
     if (age != null && age.length < 1) {
       $errors.add(
         ValidationError(
@@ -247,7 +211,87 @@ extension $UserValidasi on User {
         ),
       );
     }
+    final $fail = ({required String message, List<String> path = const []}) {
+      $errors.add(
+        ValidationError(
+          rule: 'Refine',
+          message: message,
+          path: path.isEmpty ? null : path,
+        ),
+      );
+    };
+    emailDoesNotStartWithName($fail, name: name, email: email);
+    if ($errors.isNotEmpty) {
+      return ValidasiResult(errors: $errors, isValid: false);
+    }
+    return ValidasiResult(errors: const [], isValid: true, data: this);
+  }
 
+  Future<ValidasiResult<User>> validateAsync() async {
+    final $errors = <ValidationError>[];
+    // Field: name
+    if (name != null && name.length < 2) {
+      $errors.add(
+        ValidationError(
+          rule: 'MinLength',
+          message: 'Minimum length is 2 characters',
+          details: {'length': '2'},
+          path: ['name'],
+        ),
+      );
+    }
+    if (name != null && name.length > 100) {
+      $errors.add(
+        ValidationError(
+          rule: 'MaxLength',
+          message: 'Maximum length is 100 characters',
+          details: {'length': '100'},
+          path: ['name'],
+        ),
+      );
+    }
+    // Field: email
+    if (email != null && email.length < 3) {
+      $errors.add(
+        ValidationError(
+          rule: 'MinLength',
+          message: 'Minimum length is 3 characters',
+          details: {'length': '3'},
+          path: ['email'],
+        ),
+      );
+    }
+    if (email != null && email.length > 100) {
+      $errors.add(
+        ValidationError(
+          rule: 'MaxLength',
+          message: 'Maximum length is 100 characters',
+          details: {'length': '100'},
+          path: ['email'],
+        ),
+      );
+    }
+    // Field: age
+    if (age != null && age.length < 1) {
+      $errors.add(
+        ValidationError(
+          rule: 'MinLength',
+          message: 'Minimum length is 1 characters',
+          details: {'length': '1'},
+          path: ['age'],
+        ),
+      );
+    }
+    final $fail = ({required String message, List<String> path = const []}) {
+      $errors.add(
+        ValidationError(
+          rule: 'Refine',
+          message: message,
+          path: path.isEmpty ? null : path,
+        ),
+      );
+    };
+    emailDoesNotStartWithName($fail, name: name, email: email);
     if ($errors.isNotEmpty) {
       return ValidasiResult(errors: $errors, isValid: false);
     }
@@ -256,5 +300,9 @@ extension $UserValidasi on User {
 
   ValidasiResult<V> validateField<V>(UserFields<V> field) {
     return field.validate(field.extract(this));
+  }
+
+  Future<ValidasiResult<V>> validateFieldAsync<V>(UserFields<V> field) async {
+    return field.validateAsync(field.extract(this));
   }
 }